Split L1 lifecycle histograms into separate L1LifecycleSubscriber
