+++ lmcache/v1/mp_observability/config.py
@@ -206,6 +206,7 @@ def init_observability(obs_config: ObservabilityConfig) -> EventBus:
         # First Party
         from lmcache.v1.mp_observability.subscribers.metrics import (
             L0LifecycleSubscriber,
+            L1LifecycleSubscriber,
             L1MetricsSubscriber,
             L2MetricsSubscriber,
             SMMetricsSubscriber,
@@ -214,6 +215,7 @@ def init_observability(obs_config: ObservabilityConfig) -> EventBus:
         sample_rate = obs_config.metrics_sample_rate
         bus.register_subscriber(L0LifecycleSubscriber(sample_rate=sample_rate))
         bus.register_subscriber(L1MetricsSubscriber())
+        bus.register_subscriber(L1LifecycleSubscriber(sample_rate=sample_rate))
         bus.register_subscriber(L2MetricsSubscriber())
         bus.register_subscriber(SMMetricsSubscriber())
