+++ lmcache/v1/mp_observability/subscribers/__init__.py
@@ -8,6 +8,7 @@ from lmcache.v1.mp_observability.subscribers.logging import (
 )
 from lmcache.v1.mp_observability.subscribers.metrics import (
     L0LifecycleSubscriber,
+    L1LifecycleSubscriber,
     L1MetricsSubscriber,
     SMMetricsSubscriber,
 )
@@ -17,6 +18,7 @@ from lmcache.v1.mp_observability.subscribers.tracing import (
 
 __all__ = [
     "L0LifecycleSubscriber",
+    "L1LifecycleSubscriber",
     "L1LoggingSubscriber",
     "L1MetricsSubscriber",
     "MPServerLoggingSubscriber",
