+++ lmcache/v1/mp_observability/subscribers/metrics/__init__.py
@@ -5,11 +5,15 @@ from lmcache.v1.mp_observability.subscribers.metrics.l0_lifecycle import (
     L0LifecycleSubscriber,
 )
 from lmcache.v1.mp_observability.subscribers.metrics.l1 import L1MetricsSubscriber
+from lmcache.v1.mp_observability.subscribers.metrics.l1_lifecycle import (
+    L1LifecycleSubscriber,
+)
 from lmcache.v1.mp_observability.subscribers.metrics.l2 import L2MetricsSubscriber
 from lmcache.v1.mp_observability.subscribers.metrics.sm import SMMetricsSubscriber
 
 __all__ = [
     "L0LifecycleSubscriber",
+    "L1LifecycleSubscriber",
     "L1MetricsSubscriber",
     "L2MetricsSubscriber",
     "SMMetricsSubscriber",
